+++ src/models/csrf_token.rs
@@ -3,10 +3,10 @@ use sqlx;
 
 #[derive(sqlx::FromRow, Serialize, Deserialize, Clone)]
 pub struct CsrfToken {
-    pub id: u32,
+    pub id: u64,
     pub sid: String,
     pub token: String,
-    pub expiry: Option<i64>,
+    pub expiry: Option<u64>,
     pub redirect: Option<String>,
     pub with_chat: bool,
 }
