+++ src/models/game_item_template.rs
@@ -5,8 +5,8 @@ use sqlx::{self, MySqlPool};
 
 #[derive(sqlx::FromRow, Serialize, Deserialize, Debug, Clone)]
 pub struct GameItemTemplate {
-    pub game_item_template_id: u32,
-    pub game_template_id: u32,
+    pub game_item_template_id: u64,
+    pub game_template_id: u64,
 
     pub name: String,
     pub image: Option<String>,
