+++ src/models/game_template.rs
@@ -7,7 +7,7 @@ use sqlx::{self, MySqlPool};
 
 #[derive(sqlx::FromRow, Serialize, Deserialize, Debug, Clone)]
 pub struct GameTemplate {
-    pub game_template_id: u32,
+    pub game_template_id: u64,
     pub user_id: String,
 
     pub name: String,
